Refactor FlatList
