+++ src/flatcollections-array/FlatArray/FlatArray.T.FlatList/FlatArray.FlatList.cs
@@ -7,8 +7,6 @@ partial struct FlatArray<T>
 {
     private sealed partial class FlatList : IList<T>, IReadOnlyList<T>
     {
-        private const int MinusOne = -1;
-
         private readonly int length;
 
         private readonly T[] items;
